initial commit
The initial files establishing a sql table and a first attempt at data gathering. Initial idea was to use https://files.pushshift.io/reddit/ , but GrabadoraDeVoz will be better at targeting subreddits one at a time, as opposed to a mass snapshot of all of Reddit during a month.

GrabadoraDeVoz is meant to highlight the word usage of a subreddit over time, for now...