Storing New Comments Live
This script is now set up to store new comments from a specified subreddit as they happen into a postgresql database. The script has been structured to call a function that takes a string input of the subreddit. The created UTC time code, the subreddit, and the body of text are all stored into the database. An arbitrary sample limiter has been placed on the stream for right now as I establish the modeling scripts.

Certain security information has been redacted or replaced with placeholders.