Retrieval and Cleaning Scripts Established
The first script's name was a misnomer, this has been changed to 1_retrieve to represent retrieve data and being the first script in the process flow. A new 2_clean has been created to represent that we are actually moving onto data cleaning now. As usual, security info has been redacted